VercelFlask/requirements.txt
Flask>=2.1.0
flasgger>=0.9.5
jsonschema>=4.1.0
urllib3>=1.26.7
requests>=2.26.0
PyYAML>=6.0